fix: resolve backend-size violations for factory/config/init cutover
Move factory/config/init wiring out of root_work.go into root_factory.go and split parity parsing-case builders so make backend-size passes.
